Create Solution.java
Simple function that, given a string s containing just the characters '(', ')', '{', '}', '[' and ']', determines if the input string is valid. An input string is valid if each open bracket is closed by the same type of brackets in the correct order and each closed bracket has a corresponding open bracket of the same type.